% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_probs_barplot.R
\name{assign_probs_barplot}
\alias{assign_probs_barplot}
\title{Make DAPC assignment probability barplots}
\usage{
assign_probs_barplot(
  wd,
  clumpp.wd,
  sample.names,
  sample.plot.groups = NULL,
  sample.plot.groups.order = NULL,
  sample.order = NULL,
  best.perc.var,
  best.model.number,
  plot.type,
  plot.width,
  plot.height,
  colors,
  border.color,
  cex.axis = 0.3,
  clust.method = "kmeans",
  apriori = FALSE
)
}
\arguments{
\item{wd}{working directory to create plots}

\item{clumpp.wd}{directory containing CLUMPP results}

\item{sample.names}{vector of sample names, same order as original data, numeric or character}

\item{sample.plot.groups}{vector of sample groups, same order as original data, numeric or character}

\item{sample.plot.groups.order}{vector of sample groups in desired order, numeric or character}

\item{best.perc.var}{which percent retained variance to plot? numeric}

\item{best.model.number}{which model number to plot? numeric}

\item{plot.type}{save plot as "pdf", "svg", or "png"}

\item{plot.width}{width of plot in inches}

\item{plot.height}{height of plot in inches}

\item{colors}{vector of colors for bar plot bars}

\item{border.color}{color for borders on bar, character}

\item{cex.axis}{font size for x axis sample names, default = 0.3}

\item{apriori}{do you wish to plot results from apriori individual assignment to species/populations/clusters, or results from assignment using using k-means clustering? TRUE or FALSE}
}
\description{
Make "structure-like" bar plots using discriminant analysis assignment probabilities for chosen model and proportion of retained variance. 
Returns plot and saves plot as file (svg or pdf). 
Plots will be named based on model selected.
}
