% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapc_clumpp.R
\name{dapc_clumpp}
\alias{dapc_clumpp}
\title{Run DAPC on data}
\usage{
dapc_clumpp(
  wd,
  CLUMPP_exe,
  data,
  n.groups,
  model.numbers,
  models,
  perc.var,
  scale = TRUE,
  center = TRUE,
  apriori = FALSE
)
}
\arguments{
\item{wd}{working directory to store CLUMPP files, results will be written to new subdirectory "CLUMPP"}

\item{CLUMPP_exe}{full path to CLUMPP executable}

\item{data}{data frame containing phyotypic data, one row per individual, one column per trait, no missing values}

\item{n.groups}{vector of the number of groups (populations, species, etc) in each delimitation model}

\item{model.numbers}{vector containing delimitaiton model numbers}

\item{models}{data frame containing species delimitation models, one row per individual, one column per model}

\item{perc.var}{vector containing cumulative percentages of variance to retain for discriminant analyses}

\item{scale}{scale PCA? Highly recommended unless you transform data prior to analysis. TRUE or FALSE}

\item{center}{center PCA? Highly recommended unless you transform data prior to analysis. TRUE or FALSE}

\item{apriori}{do you wish to use apriori individual assignment to species/populations/clusters, or assign individuals using k-means clustering? TRUE or FALSE}
}
\description{
Permutes the data, then performs performs K-means clustering, runs discriminant analysis, and generates "indfiles" and "paramfiles" for CLUMPP based on user supplied info.
Note that PCA with scale=TRUE is a correlation-based PCA, while PCA with scale=FALSE is a covariance-based PCA.
See https://aedin.github.io/PCAworkshop/articles/b_PCA.html.
These two types of PCA are appropriate for different datasets.
For example, covariance-based PCA is appropriate for morphometric data.
See https://www.palass.org/publications/newsletter/palaeomath-101/palaeomath-part-21-principal-warps-relative-warps-and-procrustes-pca.
While correlation-based PCA is appropriate for datasets with mixed variable types that are on different scales (e.g. meristic data).
See https://stats.stackexchange.com/questions/53/pca-on-correlation-or-covariance for more info.
"You tend to use the covariance matrix when the variable scales are similar and the correlation matrix when variables are on different scales.
}
