% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapc_clumpp_permuted.R
\name{dapc_clumpp_permuted}
\alias{dapc_clumpp_permuted}
\title{Run DAPC on permuted data}
\usage{
dapc_clumpp_permuted(
  wd,
  CLUMPP_exe,
  data,
  n.groups,
  model.numbers,
  models,
  perc.var,
  permutations,
  scale = TRUE,
  center = TRUE
)
}
\arguments{
\item{wd}{working directory to store CLUMPP files, results will be written to new subdirectory "CLUMPP_permuted"}

\item{CLUMPP_exe}{full path to CLUMPP executable}

\item{data}{data frame containing phyotypic data, one row per individual, one column per trait, no missing values}

\item{n.groups}{vector of the number of groups (populations, species, etc) in each delimitation model}

\item{model.numbers}{vector containing delimitaiton model numbers}

\item{models}{data frame containing species delimitation models, one row per individual, one column per model}

\item{perc.var}{vector containing cumulative percentages of variance to retain for discriminant analyses}

\item{permutations}{number of permutations to perform}

\item{scale}{scale PCA? Highly recommended unless you transform data prior to analysis. TRUE or FALSE}

\item{center}{center PCA? Highly recommended unless you transform data prior to analysis. TRUE or FALSE}
}
\description{
Permutes the data, then performs performs K-means clustering, runs discriminant analysis, and generates "indfiles" and "paramfiles" for CLUMPP based on user supplied info.
}
