% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_loading.R
\name{discriminant_loading}
\alias{discriminant_loading}
\title{DAPC loading plots}
\usage{
discriminant_loading(
  wd,
  clumpp.wd,
  best.perc.var,
  best.model.number,
  plot.type,
  plot.width,
  plot.height,
  axis,
  clust.method = "kmeans",
  apriori = FALSE
)
}
\arguments{
\item{wd}{working directory to create plots}

\item{clumpp.wd}{directory containing CLUMPP results}

\item{best.perc.var}{which percent retained variance to plot? numeric}

\item{best.model.number}{which model number to plot? numeric}

\item{plot.type}{save plot as "pdf", "svg", or "png"}

\item{plot.width}{width of plot in inches}

\item{plot.height}{height of plot in inches}

\item{axis}{which discriminant axis to examine? numeric}

\item{apriori}{do you wish to plot results from apriori individual assignment to species/populations/clusters, or results from assignment using using k-means clustering? TRUE or FALSE}
}
\description{
Make loading plot of variables on selected discriminant axis. Saves plot as file (svg or pdf).
Also returns file with variable contributions and loadings for that axis.
}
