% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_permutation_plot.r
\name{H_permutation_plot}
\alias{H_permutation_plot}
\title{Plot permuted CLUMPP results}
\usage{
H_permutation_plot(
  wd,
  clumpp.data,
  clumpp.data.permuted,
  model.numbers,
  best.perc.var,
  plot.type,
  plot.prefix,
  plot.width,
  plot.height
)
}
\arguments{
\item{wd}{directory to create plots, should contain "CLUMPP_permuted" subdirectory}

\item{clumpp.data}{data frame from read_clumpp_results function}

\item{clumpp.data.permuted}{data frame from read_clumpp_results_permuted function}

\item{model.numbers}{vector containing delimitaiton model numbers}

\item{best.perc.var}{which percent retained variance to plot? numeric}

\item{plot.type}{save plot as "pdf", "svg", or "png"}

\item{plot.prefix}{prefix for plot name, character}

\item{plot.width}{width of plot in inches}

\item{plot.height}{height of plot in inches}
}
\description{
Plot permuted CLUMPP H' results. Returns plot of obsserved H' minus mean permuted H'. Number inside bar is the H' difference, number on top of bar is p-value from permutation test.
}
