% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clumpp_results.R
\name{plot_clumpp_results}
\alias{plot_clumpp_results}
\title{Plot CLUMPP H' values}
\usage{
plot_clumpp_results(
  wd,
  clumpp.data,
  colors,
  plot.name,
  plot.type,
  plot.width,
  plot.height
)
}
\arguments{
\item{wd}{directory to create plots}

\item{clumpp.data}{data frame from read_clumpp_results function}

\item{colors}{vector of colors for percentage of retained variance}

\item{plot.name}{name for plot, character}

\item{plot.type}{save plot as "pdf", "svg", or "png"}

\item{plot.width}{width of plot in inches}

\item{plot.height}{height of plot in inches}
}
\description{
Plot CLUMPP H' values for each delimitation model and percetage of retained variance. Returns plot and saves plot as file (svg or pdf).
}
