% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discriminant_axes.R
\name{plot_discriminant_axes}
\alias{plot_discriminant_axes}
\title{Make scatter or density plots of discriminant axes}
\usage{
plot_discriminant_axes(
  wd,
  clumpp.wd,
  sample.plot.groups = NULL,
  sample.plot.groups.order = NULL,
  best.perc.var,
  best.model.number,
  plot.type,
  plot.width,
  plot.height,
  colors,
  shapes,
  x.axis = NULL,
  y.axis = NULL,
  apriori = FALSE,
  clust.method = "kmeans"
)
}
\arguments{
\item{wd}{directory to create plots}

\item{clumpp.wd}{directory containing CLUMPP results}

\item{sample.plot.groups}{vector of sample groups, same order as original data, numeric or character}

\item{sample.plot.groups.order}{vector of sample groups in desired order, numeric or character}

\item{best.perc.var}{which percent retained variance to plot? numeric}

\item{best.model.number}{which model number to plot? numeric}

\item{plot.type}{save plot as "pdf", "svg", or "png"}

\item{plot.width}{width of plot in inches}

\item{plot.height}{height of plot in inches}

\item{colors}{vector of colors for groups}

\item{shapes}{vector of shapes for K-means clusters}

\item{x.axis}{which discriminant axis to show on x axis? If only two clusters, discriminant axis 1 is automatically used. numeric}

\item{y.axis}{which discriminant axis to show on y axis? If only two clusters, this argument is not used. numeric}

\item{apriori}{do you wish to plot results from apriori individual assignment to species/populations/clusters, or results from assignment using using k-means clustering? TRUE or FALSE}

\item{sample.names}{vector of sample names, same order as original data, numeric or character}
}
\description{
Make density plot (2 clusters) or scatter plot (>2 clusters) of discriminant axes, as specified by the user. Returns plot and saves plot as file (svg or pdf).
}
